backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
supabase>=2.3.4
anthropic>=0.34.0
pypdf2>=3.0.1
sentence-transformers>=2.2.2
numpy>=1.24.3
python-dotenv>=1.0.0
pgvector>=0.2.4
psycopg2-binary>=2.9.9
pydantic>=2.5.2
torch>=2.0.0
transformers>=4.30.0